# docs/swagger.yaml
basePath: /api
consumes:
- application/json
definitions:
...
  NullString:
    description: |-
      var s NullString
      err := db.QueryRow("SELECT name FROM foo WHERE id=?", id).Scan(&s)
      ...
      if s.Valid {
      use s.String
      } else {
      NULL value
      }
    properties:
      String:
        type: string
      Valid:
        type: boolean
    title: |-
      NullString represents a string that may be null.
      NullString implements the Scanner interface so
      it can be used as a scan destination:
    type: object
    x-go-package: database/sql
  NullTime:
    description: |-
      NullTime implements the Scanner interface so
      it can be used as a scan destination, similar to NullString.
    properties:
      Time:
        format: date-time
        type: string
      Valid:
        type: boolean
    title: NullTime represents a time.Time that may be null.
    type: object
    x-go-package: database/sql
  User:
    properties:
      created_at:
        $ref: '#/definitions/NullTime'
      name:
        $ref: '#/definitions/NullString'
      password:
        type: string
        x-go-name: Password
      profile_url:
        $ref: '#/definitions/NullString'
      user_id:
        format: int32
        type: integer
        x-go-name: UserID
      username:
        type: string
        x-go-name: Username
    type: object
    x-go-package: github.com/nemo984/money-app-api/db/sqlc
...
  updateUserRequest:
    properties:
      name:
        type: string
        x-go-name: Name
      password:
        type: string
        x-go-name: Password
      profile_url:
        type: string
        x-go-name: ProfileURL
      username:
        type: string
        x-go-name: Username
    type: object
    x-go-package: github.com/nemo984/money-app-api/handler
  usernamePasswordRequest:
    properties:
      password:
        example: Hello123
        minLength: 6
        type: string
        x-go-name: Password
      username:
        example: Hello
        type: string
        x-go-name: Username
    required:
    - username
    - password
    type: object
    x-go-package: github.com/nemo984/money-app-api/handler
info:
  description: Documentation for Money Management API
  title: Money Management API
  version: 1.0.0
paths:
...
  /me:
    delete:
      description: Delete user of the auth token
      operationId: deleteUser
      responses:
        "204":
          $ref: '#/responses/noContent'
      security:
      - bearerAuth: []
      - cookieAuth: []
      tags:
      - Users
    get:
      description: Returns user of the auth token
      operationId: getUser
      responses:
        "200":
          $ref: '#/responses/userResponse'
      security:
      - bearerAuth: []
      - cookieAuth: []
      tags:
      - Users
    patch:
      description: Update user of the auth token with provided fields and returns
        the user
      operationId: updateUser
      parameters:
      - description: The fields to update for the user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/updateUserRequest'
        x-go-name: User
      responses:
        "200":
          $ref: '#/responses/userResponse'
      security:
      - bearerAuth: []
      - cookieAuth: []
      tags:
      - Users
...
  /me/picture:
    put:
      consumes:
      - multipart/form-data
      description: Update the profile picture of the user and returns the user
      operationId: updateUserPicture
      parameters:
      - description: Image file for profile picture < 2Mb
        in: formData
        name: file
        required: true
        type: file
        x-go-name: File
      responses:
        "200":
          $ref: '#/responses/userResponse'
      security:
      - bearerAuth: []
      - cookieAuth: []
      tags:
      - Users
...
  /token:
    post:
      description: Returns an auth token for the user
      operationId: loginUser
      parameters:
      - description: The budget to create
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/usernamePasswordRequest'
        x-go-name: User
      responses:
        "200":
          $ref: '#/responses/tokenResponse'
      tags:
      - Users
  /users:
    post:
      description: Returns the created user
      operationId: createUser
      parameters:
      - description: The budget to create
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/usernamePasswordRequest'
        x-go-name: User
      responses:
        "200":
          $ref: '#/responses/userResponse'
      tags:
      - Users
produces:
- application/json
responses:
...
  tokenResponse:
    description: Token
    headers:
      Body: {}
    schema:
      properties:
        token:
          type: string
          x-go-name: Token
      type: object
  userResponse:
    description: User
    headers:
      Body: {}
    schema:
      $ref: '#/definitions/User'
schemes:
- http
- ws
securityDefinitions:
  bearerAuth:
    description: Token obtained from loginUser endpoint
    in: header
    name: Authorization
    type: apiKey
  cookieAuth:
    in: cookie
    name: jwt-token
    type: apiKey
swagger: "2.0"